+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## 81.0.0
+## 81.0.0 (November 24, 2019)
 
 ### Enhancements
 - **Adverse Event Changes**